requirements.txt
fastapi
uvicorn
boto3
mangum
python-dotenv
python-multipart
requests